requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-CORS==4.0.0
Flask-Limiter==3.7.0
pymongo==4.7.3
bcrypt==4.1.3
Pillow==10.3.0
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
requests==2.32.3
itsdangerous==2.2.0
Werkzeug==3.0.3
gunicorn==22.0.0
dnspython==2.6.1
email-validator==2.1.1
python-jose==3.3.0
cryptography==42.0.8
